Updated to use the 'getline' function
This update removes the 'parse_helper()' since it's no longer needed.
Also, the tokenization is done directly as opposed to the previous
way of calling the tokenize function. We were losing useful information
as a result of tokenizing that way. This update fixes that and hopefully
handles it as expected.